pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.11",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
    "streamlit>=1.50.0",
    "trafilatura>=2.0.0",
]
